돌아는 가는 코드

+++ src/features/todoList/TodoItem.tsx
@@ -1,13 +1,15 @@
 import moment from "moment";
-import { useEffect, useMemo, useState } from "react";
-import { useAppDispatch } from "../../app/hooks";
+import { useState } from "react";
+import { useAppDispatch, useAppSelector } from "../../app/hooks";
 import { inThreeDay } from "../../utilities/date";
 import { setId, setModal } from "../slice/modalSlice";
-import { checkTask, deleteTask } from "../slice/taskSlice";
+import { addFilter } from "../slice/pageSlice";
+import { checkTask, deleteTask, selectTags } from "../slice/taskSlice";
 import styles from "./TodoItem.module.css";
 
 const TodoItem = ({ todoItem }) => {
   const dispatch = useAppDispatch();
+  const tags = useAppSelector(selectTags);
   const [description, setDescription] = useState<boolean>(false);
   const emergency = inThreeDay(todoItem.goalAt);
   const onEdit = (id: number) => () => {
@@ -16,35 +18,61 @@ const TodoItem = ({ todoItem }) => {
   };
   return (
     <div className={styles.todoBox}>
-      <input
-        type="checkbox"
-        onChange={() => {
-          dispatch(checkTask(todoItem.id));
-        }}
-        checked={todoItem.complete}
-      />
-      <div
-        onClick={() => {
-          setDescription(true);
-        }}
-        style={{ textDecoration: todoItem.complete ? "line-through" : "none" }}
-      >
-        {emergency ? "[긴급함]" : ""}
-        {todoItem.title}
-      </div>
-      <div className={styles.button}>
-        <button onClick={onEdit(todoItem.id)} className={styles.buttonItem}>
-          edit
-        </button>
-        <button
-          className={styles.buttonItem}
+      <div className={styles.todoItem}>
+        <input
+          type="checkbox"
+          onChange={() => {
+            dispatch(checkTask(todoItem.id));
+          }}
+          checked={todoItem.complete}
+        />
+        <div
           onClick={() => {
-            dispatch(deleteTask(todoItem.id));
+            setDescription(true);
+          }}
+          style={{
+            textDecoration: todoItem.complete ? "line-through" : "none",
           }}
         >
-          delete
-        </button>
+          {emergency ? "[긴급함]" : ""}
+          {todoItem.title}
+        </div>
+        <div className={styles.button}>
+          <button onClick={onEdit(todoItem.id)} className={styles.buttonItem}>
+            edit
+          </button>
+          <button
+            className={styles.buttonItem}
+            onClick={() => {
+              const result = confirm("정말로 삭제하시겠습니까?");
+              // const result = true;
+              if (result) {
+                dispatch(deleteTask(todoItem.id));
+              }
+            }}
+          >
+            delete
+          </button>
+        </div>
       </div>
+      <span>
+        {todoItem.tag.map((tag: string) => (
+          <span
+            style={{
+              color: tags[tag].color,
+              backgroundColor: tags[tag].backgroundColor,
+              boxSizing: "border-box",
+              padding: "0 0.5rem",
+              marginRight: "0.5rem",
+              borderRadius: "5px",
+            }}
+            key={tag}
+            onClick={() => dispatch(addFilter(tag))}
+          >
+            {tag}
+          </span>
+        ))}
+      </span>
       {description && (
         <div className={styles.ModalWrapper}>
           <div className={styles.ModalItem}>
@@ -60,8 +88,20 @@ const TodoItem = ({ todoItem }) => {
             <span>{moment(todoItem.createdAt).format("YYYY-MM-DD")}</span>
             <h2>Tag</h2>
             <span>
-              {todoItem.tag.map(() => (
-                <></>
+              {todoItem.tag.map((tag: string) => (
+                <span
+                  style={{
+                    color: tags[tag].color,
+                    backgroundColor: tags[tag].backgroundColor,
+                    boxSizing: "border-box",
+                    padding: "0 0.5rem",
+                    marginRight: "0.5rem",
+                    borderRadius: "5px",
+                  }}
+                  key={tag}
+                >
+                  {tag}
+                </span>
               ))}
             </span>
             <button onClick={() => setDescription(false)}>Close</button>
